# docs/swagger.yaml
basePath: /
definitions:
  models.Order:
    properties:
      email:
        type: string
      fullname:
        type: string
      id:
        type: integer
      paid:
        type: boolean
      payment:
        type: integer
      phone:
        type: string
      schedule:
        type: integer
      total:
        type: number
      user:
        type: integer
    required:
    - email
    - fullname
    - payment
    - phone
    - schedule
    type: object
  models.UserAuth:
    properties:
      email:
        type: string
      password:
        minLength: 4
        type: string
    required:
    - email
    - password
    type: object
  models.UserRegister:
    properties:
      email:
        type: string
      id:
        type: integer
      id_account:
        type: integer
      isactive:
        type: boolean
      password:
        minLength: 4
        type: string
      role:
        type: string
    required:
    - email
    - id_account
    - isactive
    - password
    - role
    type: object
host: localhost:8080
info:
  contact: {}
  description: Restful API craeted using gin for Koda Batch 3
  title: Weekly 10 Koda
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.UserAuth'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserAuth'
      summary: Login
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserRegister'
      summary: Register
      tags:
      - Authentication
  /history/{id}:
    get:
      description: Mengambil detail histori berdasarkan ID
      parameters:
      - description: ID History
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get History by ID
      tags:
      - History
  /movies/{id}:
    get:
      parameters:
      - description: Movie Detail
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Detail Movie
      tags:
      - Movies
  /movies/allmovie:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get All Movie
      tags:
      - Movies
  /movies/filter:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Filter movies
      tags:
      - Movies
  /movies/popular:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Popular movies
      tags:
      - Movies
  /movies/upcoming:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get upcoming movies
      tags:
      - Movies
  /order:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order Request
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
  /profile/{id}:
    get:
      parameters:
      - description: ID Profile
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Profile
      tags:
      - Profile
  /schedule/{id}:
    get:
      parameters:
      - description: Movie Schedule
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Schedule
      tags:
      - Schedule
  /seats/{id}:
    get:
      parameters:
      - description: ID Schedule
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get available seat
      tags:
      - Seat
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"